requirements.txt
openai>=1.0.0
pyyaml>=6.0
agentscope>=1.0.0

# 开发/测试
pytest>=8.0.0
pytest-mock>=3.12.0
